pyproject.toml
[tool.poetry]
name = "py_msfs_lcd1602"
version = "0.1.0"
description = "Python resources to show MSFS informations on a LCD1602"
authors = ["flocknroll <dosso.florent@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
smbus2 = "^0.4.3"
fastapi = "^0.105.0"
pyzmq = "^25.1.2"
pysimconnect = "^0.2.6"
requests = "^2.31.0"
uvicorn = "^0.25.0"
joserfc = "^0.9.0"
pendulum = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[tool.poetry.scripts]
lcd_server_api = "py_msfs_lcd1602.lcd_server.api:run"
lcd_server_controller = "py_msfs_lcd1602.lcd_server.controller:run"
msfs_service = "py_msfs_lcd1602.msfs_service.service:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
